convert index.js to ts
